Move variants processing to separate file
